build Fix rubocop config inheritance
Paths specified in configuration files starting with .rubocop are
relative to the directory where that file is. Paths in other config files
are relative to the current directory. This is so that paths in
config/default.yml, for example, are not relative to RuboCop's config
directory since that wouldn't work.