hotfix FIX-1655 Fix flaky PR validator
Fix flakiness in the PR validator, and output more details on the
associated commits for easier debugging.

In order for FuseTube to use the changes in this gem, roll back
Rubocop changes to prevent widespread impact on the FuseTube test
suite. Additionally fix the Rubocop version more tightly to prevent
changes to specific Cops from impacting testing in this gem and
consumers.