chore Release creation passes target branch correctly
Problem
The target branch was specified but not passed correctly to the
helper functions, resulting in ill-defined behaviour, including:
- creating releases or tags on the default branch, sometimes this
  was not the master branch, but a dev branch
- creating a new tag onto a (now-stale) commit

Solution
Pass the target_branch parameter appropriately.

Note
This highlights the problems with testing some of the fuse-dev-tools
functionality. I suggest a new repo, who's sole purpose is to act as
a test subject for the release-creation script. We can add some
meaningless commits and branches, and run the script against it, even
during local development and testing.