chore Document target branch flags
Problem
The release creation and changelog generation tools support a (new)
flag to target a specific branch instead of defaulting to HEAD, but
the README does not document this adequately.

Also, the additional bump options (minor-pre and patch-pre) are not
documented.

Solution
Document the new target branch flag.
Document the new bump options.